// frontend/app/src/types/contracts.d.ts
export interface components {
    schemas: {
// ...
        Pairing: {
            /**
             * Confirmed At
             * @default null
             */
            confirmed_at: string | null;
            /** Device Id */
            device_id: string;
            /** Doctor Display Name */
            doctor_display_name: string;
            /** Doctor Id */
            doctor_id: string;
            /** Doctor Pk */
            doctor_pk: string;
            /**
             * Is Demo
             * @default false
             */
            is_demo: boolean;
            /**
             * Peer Kind
             * @default doctor
             * @enum {string}
             */
            peer_kind: "doctor" | "buddy";
            /**
             * Status
             * @default pending
             * @enum {string}
             */
            status: "pending" | "paired" | "revoked";
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}